refactor: extract backend podcast category mock handler to separate file
