+++ backend/__tests__/mocks/handlers/podcastStatHandler.ts
@@ -1,5 +1,5 @@
 import { http, HttpHandler, HttpResponse } from "msw"
-import { CURRENT_PODCAST_STATISTICS } from "./data/podcastStatistic.js"
+import { CURRENT_PODCAST_STATISTICS } from "../data/podcastStatistic.js"
 
 export const podcastStatsHandlers: HttpHandler[] = [
   http.get("https://api.podcastindex.org/api/1.0/stats/current", () => {
